+++ week4/src/apis/memberJoin.ts
@@ -1,23 +1,16 @@
+import { isAxiosError } from "axios";
+import { JoinType } from "../types";
 import { serverAxios } from "./axios";
-type Props = {
-  authenticationId: string;
-  password: string;
-  nickname: string;
-  phone: string;
-};
+
 /** 회원가입 */
-export const memberJoin = async (props: Props) => {
-  const data = {
-    authenticationId: props.authenticationId,
-    password: props.password,
-    nickname: props.nickname,
-    phone: props.phone,
-  };
+export const memberJoin = async (props: JoinType) => {
   try {
-    const res = await serverAxios.post("/member/join", data);
-    console.log(res);
+    const res = await serverAxios.post("/member/join", props);
     return res;
   } catch (error) {
-    console.log(error);
+    if (isAxiosError(error)) alert(error.response?.data.message);
+    else {
+      console.log(error, "unknown error: memberJoin");
+    }
   }
 };
