+++ week4/src/types.ts
@@ -2,3 +2,9 @@ export interface LoginType {
   authenticationId: string;
   password: string;
 }
+export interface JoinType {
+  authenticationId: string;
+  password: string;
+  nickname: string;
+  phone: string;
+}
